Refactor navigator runtime assembly
