+++ bootstrap/navigator/runtime.py
@@ -7,11 +7,12 @@ from typing import Protocol
 from navigator.app.service.navigator_runtime import MissingAlert
 from navigator.core.telemetry import Telemetry
 from navigator.infra.di.container import AppContainer
-from navigator.presentation.bootstrap.navigator import compose
+from navigator.presentation.bootstrap.navigator import NavigatorDependencies, compose
 from navigator.presentation.navigator import Navigator
 
 from .context import BootstrapContext, scope_from_dto
 from .container import ContainerFactory
+from .inspection import NavigatorContainerSnapshot, inspect_container
 from .telemetry import TelemetryFactory, calibrate_telemetry
 
 
@@ -42,8 +43,9 @@ class ContainerRuntimeFactory(NavigatorFactory):
     async def create(self, context: BootstrapContext) -> NavigatorRuntimeBundle:
         telemetry = self._create_telemetry()
         container = self._create_container(context, telemetry)
-        self._calibrate(telemetry, container)
-        navigator = self._compose_navigator(container, context)
+        snapshot = inspect_container(container)
+        self._calibrate(telemetry, snapshot)
+        navigator = self._compose_navigator(snapshot, context)
         return NavigatorRuntimeBundle(telemetry=telemetry, container=container, navigator=navigator)
 
     def _create_telemetry(self) -> Telemetry:
@@ -53,17 +55,17 @@ class ContainerRuntimeFactory(NavigatorFactory):
         factory = ContainerFactory(telemetry, alert=self._missing_alert)
         return factory.create(context)
 
-    def _calibrate(self, telemetry: Telemetry, container: AppContainer) -> None:
-        core = container.core()
-        settings = core.settings()
-        calibrate_telemetry(telemetry, getattr(settings, "redaction", ""))
+    def _calibrate(self, telemetry: Telemetry, snapshot: NavigatorContainerSnapshot) -> None:
+        calibrate_telemetry(telemetry, snapshot.redaction)
 
-    def _compose_navigator(self, container: AppContainer, context: BootstrapContext) -> Navigator:
-        core = container.core()
+    def _compose_navigator(
+        self, snapshot: NavigatorContainerSnapshot, context: BootstrapContext
+    ) -> Navigator:
+        dependencies: NavigatorDependencies = snapshot.dependencies
         return compose(
-            container,
+            dependencies,
             scope_from_dto(context.scope),
-            missing_alert=core.alert(),
+            missing_alert=context.missing_alert,
         )
